pyproject.toml
[project]
name = "damn-cat-everyday-bot"
version = "0.1.0"
authors = [{ name = "Mat" }]
description = "Telegram bot that posts that orange cat with a new caption every day"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocron>=2.1",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot[http2]>=22.5",
]

[dependency-groups]
dev = [
    "mypy[faster-cache]>=1.19.1",
    "ruff>=0.14.10",
]

[tool.ruff]
target-version = "py313"
line-length = 100
indent-width = 4
